Create basic stand-alone app example
Clone the original example as stand-alone-example
in order to demonstrate launching an app using this
package directly on the CarPlay-/AndroidAuto-client
without having the app running on the phone.

The stand-alone-example only renders a placeholder
text on phone and contains all it's CarPlay-logic
inside hooks which listen for
- CarPlay connection changes
- state changes (via useReducer)

The stand-alone-example CarPlay-app features
a TabTemplate as root template containing
a top-level ListTemplate with two browsable items.
Selecting a top-level item pushes a new
(sub-level) ListTemplate onto the CarPlay-stack,
which contains two non-browsable items.
Selecting a sub-level item presents a CarPlay
modal.